+++ src/ruleSchema/ruleSchema.ts
@@ -1,0 +1,6 @@
+/**
+ * The rule for schema node
+ */
+export class SchemaNodeRule {
+    
+}
